[ AMSD-117 ] Implement the authentication and authorization state in the passenger baggage and airlines component

+++ src/Components/Baggage/BaggageList.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useRouteLoaderData } from 'react-router-dom';
 import DataTable from '../../util/DataTable'; // Adjust the path based on your directory structure
 import { Container, Box, Typography, IconButton, Button } from '@mui/material';
 import EditIcon from '@mui/icons-material/Edit';
@@ -8,6 +8,7 @@ import { apiService } from '../../util/apiService';
 
 export default function BaggageList() {
   const [baggages, setBaggages] = useState([]);
+  const auth = useRouteLoaderData('root');
 
   useEffect(() => {
     loadBaggages();
@@ -34,7 +35,20 @@ export default function BaggageList() {
   };
 
   const columns = [
-    { field: 'baggageId', headerName: 'ID', width: 100 },
+    {
+      field: 'baggageId',
+      headerName: 'ID',
+      width: 100,
+      renderCell: (params) => (
+        auth ? (
+          <Link to={`${params.row.baggageId}`} style={{ textDecoration: 'none', color: 'inherit' }}>
+            {params.row.baggageId}
+          </Link>
+        ) : (
+          <span>{params.row.baggageId}</span>
+        )
+      )
+    },
     {
       field: 'passenger',
       headerName: 'Passenger',
@@ -44,10 +58,12 @@ export default function BaggageList() {
         const passengerDisplay = passenger 
           ? `${passenger.passengerId} - ${passenger.name}`
           : 'N/A';
-        return (
-          <Link to={`/baggage-passenger-details/${passenger ? passenger.passengerId : ''}`} style={{ textDecoration: 'none', color: 'inherit' }}>
+        return auth ? (
+          <Link to={`../passengers/${passenger.passengerId}`} style={{ textDecoration: 'none', color: 'inherit' }}>
             {passengerDisplay}
           </Link>
+        ) : (
+          <span>{passengerDisplay}</span>
         );
       }
     },
@@ -57,53 +73,57 @@ export default function BaggageList() {
       width: 200,
       valueGetter: (params) => {
         const flight = params.row.flight;
-        console.log('Flight:', flight); // Debug log
         return flight 
           ? `${flight.flightId} - ${flight.flightNumber}`
           : 'N/A';
       }
     },
-    { field: 'weight', headerName: 'Weight', width: 150 },
-    {
-      field: 'edit',
-      headerName: 'Edit',
-      width: 100,
-      renderCell: (params) => (
-        <Link to={`/edit-baggage/${params.row.baggageId}`} style={{ textDecoration: 'none' }}>
-          <IconButton color="primary">
-            <EditIcon />
-          </IconButton>
-        </Link>
-      )
-    },
-    {
-      field: 'delete',
-      headerName: 'Delete',
-      width: 100,
-      renderCell: (params) => (
-        <IconButton
-          color="secondary"
-          onClick={() => handleDeleteBaggage(params.row.baggageId)}
-        >
-          <DeleteIcon />
-        </IconButton>
-      )
-    }
+    { field: 'weight', headerName: 'Weight', width: 150 }
   ];
 
+  if (auth) {
+    columns.push(
+      {
+        field: 'edit',
+        headerName: 'Edit',
+        width: 100,
+        renderCell: (params) => (
+          <Link to={`${params.row.baggageId}/edit`} style={{ textDecoration: 'none' }}>
+            <IconButton color="primary">
+              <EditIcon />
+            </IconButton>
+          </Link>
+        )
+      },
+      {
+        field: 'delete',
+        headerName: 'Delete',
+        width: 100,
+        renderCell: (params) => (
+          <IconButton
+            color="secondary"
+            onClick={() => handleDeleteBaggage(params.row.baggageId)}
+          >
+            <DeleteIcon />
+          </IconButton>
+        )
+      }
+    );
+  }
+
   return (
-    <Container
-      className='home'
-      maxWidth="lg">
+    <Container maxWidth="lg">
       <Box textAlign="center">
         <Typography variant="h4" component="h1" gutterBottom>
           Baggages
         </Typography>
-        <Link to="/add-baggage" style={{ textDecoration: 'none' }}>
-          <Button variant="contained" color="primary" style={{ marginRight: '10px' }}>
-            Add Baggage
-          </Button>
-        </Link>
+        {auth && (
+          <Link to="new" style={{ textDecoration: 'none' }}>
+            <Button variant="contained" color="primary" style={{ marginRight: '10px' }}>
+              Add Baggage
+            </Button>
+          </Link>
+        )}
       </Box>
       <Box sx={{ display: 'flex', justifyContent: 'center', marginBottom: 3 }}>
         <Box sx={{ width: '100%', maxWidth: 1000 }}>
